Update DynamicHlsService.cs
Changing command line arguments to ffmpeg to begin transcoding.  Existing command line generates segments that are approximate length to the "length" parameter passed in.  In many cases this would cause a mismatch between the expected number of segments, and the actual number of segments.  Updated options yield segments that more strictly conform to "segment length" parameter, causing the number of segments to match the number expected.