Smart-Match: Added feature to remember/persist series matching in manual organization dialog
When a filename cannot be auto-matched to an existing series name, the organization must be performed manually.
Unfortunately not just once, but again and again for each episode coming in.
This change proposes a simple but solid method to optionally persist the matching condition from within the manual organization dialog.
This approach will make Emby "learn" how to organize files in the future without user interaction.