Reworked library refresh implementation
The LibraryMonitor was implemented in a way where a single locked file
path queued for refresh kept blocking all other queued paths from being
refreshed.

An example why this was really bad: I have configured my WMC "Recorded
TV" folder as a library folder. Regardless of the time of day - in most
cases there is at least a single tv recording active. The LibraryMonitor
detected the new recording and added it to an internal queue for
processing. But processing does not start as long as even just a single
file in the queue is still locked (=being written).

After this change, all non-locked files are identified and processed
immediately while only locked files are deferred for later processing.