Fixed bug in librarybrowser.js
Updated getPosterViewItemHtml to fix a bug in which a background image
was not displayed for certain movies in the "Latest Media" section of
the web interface.  More specifically, a movie that was part of a
collection (i.e. has a parent item) AND did not have a Thumb image
would not display any image.  It would attempt to get a Thumb image
from it’s parent (i.e. the collection) but would fail because the
function ApiClient.getThumbImageUrl is seriously broken.

Even if, ApiClient.getThumbImageUrl was NOT broken, it makes more sense
to show a background image for a movie than it does to show the Thumb
of a collection in which it resides.  So the precedence for these two
cases has been swapped.

The ApiClient.getThumbImageUrl function is still broken.  At the time
of this commit is it not clear what situations exercise the
‘options.preferThumb && item.ParentThumbItemId && options.inheritThumb
!== false’ case.  Since that code is broken anyhow, I assume it is not
a very important case.