EpisodeFileOrganizer: Use Exceptions instead of null return values (Reworked)
Previously some methods were just returning null  or empty values in
case of encountered errors; as a consequence, the actual reason for
failure was never propagated to the UI which instead just displayed a
generic message like "Unable to sort xxx because target path could not
be determined."
After this change, the actual reason for failure will be saved to the
auto-organize log or displayed in the UI (when completing the organize
dialog).

This information is very important for the user. Examples are "No
permission", "Target folder not available", "Disk full", etc..