EpisodeFileOrganizer: Improve error handling (3rd alternate - minmal version)
Previously some methods were just returning null or empty values in case
of encountered errors; as a consequence, the actual reason for failure
was never written to the auto-organize log.
Instead, only a generic message like "Unable to sort xxx because target
path could not be determined." was displayed.

After this change, the actual reason for failure will be saved to the
auto-organize log or displayed in the UI (when completing the organize
dialog).

This information is very important for the user. Examples are "No
permission", "Target folder not available", "Disk full", etc..