Various Improvements to Support Development, Troubleshooting and Message Filtering
Motivation was to provide a transparent way in which to apply custom
filters to messages being written to the log. This will assist in
debugging/troubleshooting, as the log can be swamped by debug messages
that are not always of interest or relevant to the task at hand. It will
also allow a custom filters to be setup for those users that wish to
maintain their existing power saving/HDD hibernation capabilities. A
mechanism has been added to address the issue of debug messages being
lost during startup, a single change in the calling code is required to
take advantage of the improvement. Debugging support, that is
independent of the logger itself, has also been added to assist with
current and future development of the logging facility. The library is
fully backwards compatible and all new functionality is implemented in
such a way that it will not throw any exceptions.