Respect a given runtime for live tv and recordings
If a tv backend provides this information already, than it's doing it for a good reason.
There's no reason to overwrite the runtime in such a case with one determined by ffprobe.
A particular use case is an inprogress recording with a file as input. This results in wrong durations in the players if ffprobe's RunTimeTicks are used.